build(gomodule): update package github.com/genshen/cmds to avoid to call exist(2) in subcommand help
In preview code, program will exit(2) when runing in terminal like `wssocks client --help`.
In this commit, it will exit(0) with help message printed.

BREAKING CHANGE:
We also update the minimal go building version to 1.13 due to the usage of feature `errors.Is` to
handle results of command line parsing.
go version less then 1.12 (including 1.12) is not supported from this commit.